update yatvgrabber
git-svn-id: https://yatvgrabber.googlecode.com/svn/trunk@213 1b0409f8-bb2b-11dd-a80a-757f205f25d6

+++ yatvgrabber/grab-tvtv.sh
@@ -4,21 +4,17 @@ XMLTV_FILE="/tmp/tvtv.xml"
 LOG_FILE="/tmp/grab_tvtv.log"
 SOURCE_IDS="1"
 
+set -e
+
+touch $LOG_FILE
 chmod 666 $LOG_FILE
 date >> $LOG_FILE
 
 #remove the old file
-if [ -e $XMLTV_FILE ]; then
-    mv -f $XMLTV_FILE ${XMLTV_FILE}.old >> $LOG_FILE
-fi
+[ -e $XMLTV_FILE ] && mv -f $XMLTV_FILE ${XMLTV_FILE}.old >> $LOG_FILE
 
 #get the data
 nice /usr/bin/python /usr/local/sbin/grabtv.py --outputfile $XMLTV_FILE >> $LOG_FILE
-retValue=$?
-if [ "$retValue" -ne "0" ]; then
-    echo "aborting, grabtv.py returned $retValue" >> $LOG_FILE
-    exit $retValue
-fi
 
 # add the data to the db
 if [ -e $XMLTV_FILE ]; then
